% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maeView.R
\name{maeView}
\alias{maeView}
\title{Title MultiAssayExperiment viewer}
\usage{
maeView(mae)
}
\arguments{
\item{mae}{A MultiAssayExperiment object}
}
\description{
Title MultiAssayExperiment viewer
}
\examples{
maeView(miniACC)
}
